Switch to 11ty 3.0.0-alpha
This lets us remove the nasty CJS-to-ESM shim.

It also breaks the sass plugin which hasn't been updated in a
while. After trying a bunch of plugins I ended up just adding
a custom extension to process the sass myself.

This also fixes fonts not working properly, because the old
stylesheet ended up in `static` for some reason but all the
other static assets were lifted into the root folder