Add a published flag to everything
In order to actually make stuff in the admin site, it needs to
be saved with an ID before being visible on the main website.
Accomplish this by adding a published flag to all content
types, and avoid showing unpublished material in any lists,
related links, etc. You can still access it by URL directly.

Also add a way to make new (unpublished) posts in the admin
UI.